+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* add install --distro to install distros
+* add install --pick to pick a version to install
+* add install --pick-distro to pick a distro to install
+* add install --force to force install a version
+
 ## 0.0.1
 
 * add `nvim-manager version` command to tell wht the current version of nvim-manager is